size() should return off_t
FileHandle size returns off_t instead of size_t

+++ platform/LocalFileSystem.h
@@ -46,7 +46,7 @@ public:
 
     virtual int sync();
 
-    virtual size_t size();
+    virtual off_t size();
 
 protected:
     virtual void lock();
